# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurante API",
        "description": "Documentación del backend (Laravel + Swagger).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Lista de pedidos",
                "operationId": "PedidosIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear pedido con items",
                "operationId": "acf9c44ffbceb280fd290ecbb5f5f3f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_cliente",
                                    "items"
                                ],
                                "properties": {
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mesa": {
                                        "type": "string",
                                        "example": "A3"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id_menu_item": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "nombre_producto": {
                                                    "type": "string",
                                                    "example": "Limonada"
                                                },
                                                "precio": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5.5
                                                },
                                                "categoria": {
                                                    "type": "string",
                                                    "example": "bebida"
                                                },
                                                "cantidad": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "sin azúcar"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Ver pedido por ID",
                "operationId": "PedidosShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualizar pedido (mesa/estado)",
                "operationId": "3da8de6967b1e89d1d5a0cf5988c04e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mesa": {
                                        "type": "string"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "en_entrega",
                                            "listo",
                                            "entregado",
                                            "cancelado"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Eliminar pedido",
                "operationId": "c0cc39e3f79740c3fe58d2268a321d0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/pedidos/{id}/detalle": {
            "get": {
                "tags": [
                    "Pedidos - Detalle"
                ],
                "summary": "Detalle del pedido",
                "operationId": "b5decb9dd04d9964b0fe67a51601d794",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
...
}